[完美][master][pep-150] PEP-150 实现PEP平台权限相关接口（用户组接口部分） (#155)
* auth relations

* user module design

* 增加逻辑删除

* 增加逻辑删除

* auth's function module include menu & resource

* remove sequenceNumber from resource entity

* remove unused import

* Merge remote-tracking branch 'remotes/王贺/auth' into local-王贺auth

# Conflicts:
#	subprojects/pep-auth/pep-auth-common/src/main/java/com/proper/enterprise/platform/auth/common/controller/UsersController.java

* Merge remote-tracking branch 'remotes/王贺/auth' into local-王贺auth

# Conflicts:
#	subprojects/pep-auth/pep-auth-common/src/main/java/com/proper/enterprise/platform/auth/common/controller/UsersController.java

* refactor menu controller

* fix

* update to 0.3.4-SNAPSHOT

* merged

* 去掉不用的导入

* 现阶段完成了实现用户的工作

* 添加异常抛出

* 去掉抛出异常

* 完善后台各层代码

* 增加测试

* 增加自定义sql语句

* merged

* 实现功能细节

* 增加测试用例

* build

* 增加了判断是否存在循环继承的代码

* Build Success

* Build Success

* Merge remote-tracking branch 'remotes/公司远程/master' into local-王贺auth-PEP-140

# Conflicts:
#	subprojects/pep-auth/pep-auth-api/src/main/java/com/proper/enterprise/platform/api/auth/service/RoleService.java
#	subprojects/pep-auth/pep-auth-api/src/main/java/com/proper/enterprise/platform/api/auth/service/UserService.java
#	subprojects/pep-auth/pep-auth-common/src/main/java/com/proper/enterprise/platform/auth/common/controller/UsersController.java
#	subprojects/pep-auth/pep-auth-common/src/main/java/com/proper/enterprise/platform/auth/common/service/impl/AbstractUserServiceImpl.java
#	subprojects/pep-auth/pep-auth-common/src/main/java/com/proper/enterprise/platform/auth/common/service/impl/RoleServiceImpl.java
#	subprojects/pep-auth/pep-auth-common/src/test/groovy/com/proper/enterprise/platform/auth/common/controller/UsersControllerTest.groovy

* 增加属性，分离用户的角色、用户组信息到新的接口，等待新接口完成后继续。同时增加了对权限的判断，超级管理员等细节处的判断。

* 完善权限代码结构，增加资源权限判断、角色、用户、用户组，各级别的权限判断。

* 增加测试用例

* 增加权限检查

* 添加@JsonIgnore,阻止转换出现级联循环

* 修改错误

* 增加异常处理enable处理

* 增加测试用例

* 增加测试用例，提高覆盖率

* 增加测试用例，提高覆盖率

* 增加去掉的接口

* 增加测试类

* 增加覆盖率

+++ subprojects/pep-auth/pep-auth-common/src/main/java/com/proper/enterprise/platform/auth/common/controller/UserGroupController.java
@@ -4,7 +4,9 @@ import com.proper.enterprise.platform.api.auth.model.Role;
 import com.proper.enterprise.platform.api.auth.model.User;
 import com.proper.enterprise.platform.api.auth.model.UserGroup;
 import com.proper.enterprise.platform.api.auth.service.UserGroupService;
+import com.proper.enterprise.platform.api.auth.service.UserService;
 import com.proper.enterprise.platform.core.controller.BaseController;
+import com.proper.enterprise.platform.core.utils.CollectionUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -19,14 +21,19 @@ public class UserGroupController extends BaseController {
     @Autowired
     private UserGroupService service;
 
+    @Autowired
+    private UserService userService;
+
     @GetMapping
     public ResponseEntity<Collection<? extends UserGroup>> getGroups(String name, String description, String enable) {
+        userService.checkPermission("/auth/user-groups", RequestMethod.GET);
         return responseOfGet(service.getGroups(name, description, enable));
     }
 
     @SuppressWarnings("unchecked")
     @PutMapping
     public ResponseEntity<Collection<? extends UserGroup>> updateEnable(@RequestBody Map<String, Object> reqMap) {
+        userService.checkPermission("/auth/user-groups", RequestMethod.PUT);
         Collection<String> idList = (Collection<String>)reqMap.get("ids");
         boolean enable = (boolean) reqMap.get("enable");
         // TODO 具体实现
@@ -35,22 +42,25 @@ public class UserGroupController extends BaseController {
 
     @PostMapping
     public ResponseEntity<UserGroup> create(@RequestBody Map<String, Object> reqMap) {
+        userService.checkPermission("/auth/user-groups", RequestMethod.POST);
         return responseOfPost(service.save(reqMap));
     }
 
     @DeleteMapping
     public ResponseEntity deleteGroups(@RequestParam String ids) {
-        // TODO 具体实现
+        userService.checkPermission("/auth/user-groups", RequestMethod.DELETE);
         return responseOfDelete(service.deleteByIds(ids));
     }
 
     @GetMapping("/{id}")
     public ResponseEntity<UserGroup> get(@PathVariable String id) {
+        userService.checkPermission("/auth/user-groups/" + id, RequestMethod.GET);
         return responseOfGet(service.get(id));
     }
 
     @PutMapping("/{id}")
     public ResponseEntity<UserGroup> update(@PathVariable String id, @RequestBody Map<String, Object> reqMap) {
+        userService.checkPermission("/auth/user-groups/" + id, RequestMethod.PUT);
         UserGroup group = service.get(id);
         if (group != null) {
             reqMap.put("id", id);
@@ -61,9 +71,9 @@ public class UserGroupController extends BaseController {
 
     @DeleteMapping("/{id}")
     public ResponseEntity delete(@PathVariable String id) {
-        // TODO 具体实现
+        userService.checkPermission("/auth/user-groups/" + id, RequestMethod.DELETE);
         UserGroup group = service.get(id);
-        if (group != null) {
+        if (group != null && CollectionUtil.isEmpty(group.getUsers())) {
             service.delete(group);
             return responseOfDelete(true);
         }
@@ -72,11 +82,13 @@ public class UserGroupController extends BaseController {
 
     @PostMapping("/{id}/role/{roleId}")
     public ResponseEntity<UserGroup> addUserGroupRole(@PathVariable String id, @PathVariable String roleId) {
+        userService.checkPermission("/auth/user-groups/" + id + "/role/" + roleId, RequestMethod.POST);
         return responseOfPost(service.saveUserGroupRole(id, roleId));
     }
 
     @DeleteMapping("/{id}/role/{roleId}")
     public ResponseEntity deleteUserGroupRole(@PathVariable String id, @PathVariable String roleId) {
+        userService.checkPermission("/auth/user-groups/" + id + "/role/" + roleId, RequestMethod.DELETE);
         return responseOfDelete(service.deleteUserGroupRole(id, roleId) != null);
     }
 
@@ -95,7 +107,7 @@ public class UserGroupController extends BaseController {
      */
     @PostMapping(path = "/{groupId}/user/{userId}")
     public ResponseEntity<UserGroup> addUserGroup(@PathVariable String groupId, @PathVariable String userId) {
-        // TODO 具体业务实现
+        userService.checkPermission("/auth/user-groups/" + groupId + "/user/" + userId, RequestMethod.POST);
         return responseOfPost(service.addGroupUser(groupId, userId));
     }
 
@@ -108,7 +120,7 @@ public class UserGroupController extends BaseController {
      */
     @DeleteMapping(path = "/{groupId}/user/{userId}")
     public ResponseEntity deleteUserGroup(@PathVariable String groupId, @PathVariable String userId) {
-        // TODO 具体业务实现
+        userService.checkPermission("/auth/user-groups/" + groupId + "/user/" + userId, RequestMethod.DELETE);
         return responseOfDelete(service.deleteGroupUser(groupId, userId) != null);
     }
